refactor: collapsed filter sidebars
